requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
faster-whisper==0.6.0
argparse==1.4.0
python-multipart==0.0.6
python-dotenv==1.0.0
nvidia-ml-py3